+++ Graphics/Sprite.h
@@ -26,18 +26,18 @@ namespace Graphics
 	class Sprite
 	{
 	public:
-		Sprite(node, vector2d<int32_t>);
+		Sprite(node, vector2d<int16_t>);
 		Sprite(node);
 		~Sprite();
 
 		void update();
 		void draw(const DrawArgument&, float) const;
-		vector2d<int32_t> getorigin() const;
-		vector2d<int32_t> getdimensions() const;
+		vector2d<int16_t> getorigin() const;
+		vector2d<int16_t> getdimensions() const;
 
 	private:
 		Animation animation;
-		vector2d<int32_t> position;
+		vector2d<int16_t> position;
 	};
 }
