+++ IO/Components/Nametag.cpp
@@ -20,17 +20,17 @@ namespace IO
 
 	Nametag::~Nametag() {}
 
-	void Nametag::draw(vector2d<int32_t> position) const
+	void Nametag::draw(vector2d<int16_t> position) const
 	{
 		if (tag.count(selected))
 		{
 			int16_t width = name.getdimensions().x();
-			vector2d<int32_t> startpos = position - vector2d<int32_t>(8 + width / 2, -3);
+			vector2d<int16_t> startpos = position - vector2d<int16_t>(8 + width / 2, -3);
 
 			using::Graphics::DrawArgument;
 			tag.at(selected).at(0).draw(DrawArgument(startpos));
-			tag.at(selected).at(1).draw(DrawArgument(startpos + vector2d<int32_t>(8, 0), vector2d<int32_t>(width, 0)));
-			tag.at(selected).at(2).draw(DrawArgument(startpos + vector2d<int32_t>(width + 8, 0)));
+			tag.at(selected).at(1).draw(DrawArgument(startpos + vector2d<int16_t>(8, 0), vector2d<int16_t>(width, 0)));
+			tag.at(selected).at(2).draw(DrawArgument(startpos + vector2d<int16_t>(width + 8, 0)));
 			name.draw(position);
 		}
 	}
